Rename useUserPreferredEdition file to match export
