+++ dotcom-rendering/src/lib/useEditionSwitcherBanner.ts
